+++ CHANGELOG.md
@@ -8,6 +8,22 @@ Versioning](https://semver.org/spec/v2.0.0.html).
 
 <!-- ## [Unreleased] -->
 
+## [0.9.5] - 2023-02-06
+
+### Changed
+
+- Better attribute socket errors, and don't crash when a socket is closed
+  unexpectedly.
+
+### Fixed
+
+- Fixed infinite loops that could occur in watch mode when a script failed, but
+  still emitted output that was configured as the input files for another
+  script.
+
+- Don't clear the console or emit "no-op" style log messages in watch mode for
+  iterations that don't do anything useful.
+
 ## [0.9.4] - 2023-01-30
 
 ### Changed